chore(web): add basic functionality to create a new snippet from client
For some unknown reason, I had to change the default status of the post
handler from `StatusSeeOther` to `StatusCreated`. Otherwise, a
successful POST request would result in an immediate redirect, such that
the response from that POST request would be equivalent to a GET request
with the id of the newly created snippet – without actually redirecting
on the client. (Maybe I could use something like `redirect: "manual"`
for that? Besides, I read that the proper status code in this scenario
is actually `201`.)